+++ server/tutorweb_quizdb/stage/__init__.py
@@ -1,4 +1,5 @@
 def includeme(config):
+    config.include('tutorweb_quizdb.stage.dataframe')
     config.include('tutorweb_quizdb.stage.index')
     config.include('tutorweb_quizdb.stage.material')
     config.include('tutorweb_quizdb.stage.ug_extensions')